modified BoardController => BoardStore / changed datasource of board in board.test.js
